requirements.txt
pandas
konlpy
sanic
sanic_session
python-socketio